backup_management/requirements.txt
boto3~=1.35.70
python-dotenv~=1.0.1